renamed ansi_codes to something more meaningful
