+++ shell/src/output/textStylesAndColors.h
